+++ tests/codeception/acceptance/template/CreateTemplatePageCest.php
@@ -88,7 +88,6 @@ class CreateTemplatePageCest
     {
         $I->click('Add Element');
         $I->wait(1);
-        $I->scrollToBottom();
         $I->click($type);
         $I->waitForElementVisible('#templateelement-name');
     }